requirements.txt
amqp==2.6.1
asgiref==3.3.4
billiard==3.6.4.0
boto3==1.17.97
botocore==1.20.97
cachetools==4.2.2
celery==4.4.1
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.4.2
Django==3.2.4
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-cors-headers==3.7.0
django-extensions==3.1.3
django-storages==1.11.1
django-timezone-field==4.1.2
djangorestframework==3.12.4
flower==0.9.7
google-api-core==1.30.0
google-api-python-client==2.9.0
google-auth==1.31.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
humanize==3.9.0
idna==2.10
importlib-metadata==4.5.0
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
kombu==4.6.11
MarkupSafe==2.0.1
packaging==20.9
prometheus-client==0.8.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycurl==7.43.0.2
PyJWT==1.7.1
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
raven==6.10.0
redis==3.5.3
requests==2.25.1
rsa==4.7.2
s3transfer==0.4.2
Shapely==1.7.1
six==1.16.0
sqlparse==0.4.1
stripe==2.58.0
tornado==6.1
twilio==6.32.0
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
vine==1.3.0
zipp==3.4.1
